issue-49: Cleanup + move notification logic to boot + custom backend url
